+++ experimental/a/assyrian_audo/source/help/assyrian_audo.php
@@ -1,0 +1,24 @@
+<?php 
+  $pagename = 'Assyrian Audo Keyboard Help';
+  $pagetitle = $pagename;
+  require_once('header.php');
+?>
+
+<p>
+    Assyrian Audo keyboard is designed for the Assyrian language.  This keyboard is a non-phonetic layout, and it is packaged with the Ramsina font.
+</p>
+
+<h2>Desktop/Tablet Keyboard Layout</h2>
+<div id='osk' data-states='default shift'>
+</div>
+
+<h2>Mobile/Phone Keyboard Layout</h2>
+<p>Due to the size and number of keys, some characters are hidden in the long press. 
+	Press and hold on the key with a little dot on the top right to reveal and use them.</p>
+
+<div id='osk-phone' data-states='default diacritics numeric'>
+</div>
+
+  <h3>Longpress</h3>
+	<p>On mobile phone, keys with a little dot on the top right can be pressed and held for more keys.</p>
+	<p><a href="phone_longpress.png"><img class="keyboard" src="phone_longpress.png" alt="Longpress state" /></a></p>