+++ release/a/arabic_flick/source/help/arabic_flick.php
@@ -1,0 +1,30 @@
+<?php 
+  $pagename = 'Arabic Flick Keyboard Help';
+  $pagetitle = $pagename;
+  require_once('header.php');
+?>
+
+
+<p>
+   Flick style keyboard for Arabic, like the T9 Style keyboard for Japanese. 
+</p>
+<p>
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift rightalt rightalt-shift'>
+</div>
+</p>
+<p>
+<h2>Mobile Keyboard Layout</h2>
+<div id='osk-phone' data-states='default'>
+Main press is the number you see on main. Swipe to get the number written in hints, and there order is as follows: Right, Top, Left, Down. Similarly, can also be long pressed in to get the characters in the order they are written in key hints.The arrows are to move, right and left. Symbols key leads you multi-layer symbol row, while number key leads you to T9 style number row. The shadda below can be longpressed, or multitapped, or flicked for more diacritics.
+</div>
+
+<div id='osk-phone' data-states='numbers'>
+The main numbers are Arabic Numerals, and Swipe-Up are Hindi Numerals. Double tapping on numbers doesn't change the letters, as numbers are repeated many times.
+</div>
+
+<div id='osk-phone' data-states='base_symbols base_symbols_two base_symbols_three currency'>
+</div>
+
+
+  </p>