+++ experimental/h/hindi_flick/HISTORY.md
@@ -1,6 +1,15 @@
 Hindi Flick Change History
 ====================
 
+1.1.1 (2026-06-26)
+------------------
+* Two characters in Shift Stage weren't going back to no_shift stage after pressing.
+
+1.1 (2026-06-10)
+------------------
+* Assigned copyright to original author Raghav Goel
+* Added rules for a desktop layout
+
 1.0.1 (2025-09-10)
 ------------------
 * Additional handling for start-of-line scenarios