+++ experimental/t/tai_lanna_thai/HISTORY.md
@@ -1,5 +1,9 @@
-Tai Lanna (Thai Layout) Change History
+Tham Lanna (Thai Layout) Change History
 ====================
+
+1.1 (2026-06-10)
+----------------
+* Rename keyboard from "Tai Lanna" to "Tham Lanna"
 
 1.0.1 (2025-08-08)
 ----------------